Changed all size_t to int type for grade value | ex03 : Wrong form name now throw an exception | Intern.cpp : Defined operator= overload to match norm

+++ ex00/includes/Bureaucrat.hpp
@@ -8,11 +8,11 @@ class Bureaucrat
 {
 private:
   const std::string _name;
-  size_t _grade;
+  int _grade;
 
 public:
   Bureaucrat();
-  Bureaucrat(std::string name, size_t grade);
+  Bureaucrat(std::string name, int grade);
   Bureaucrat(const Bureaucrat &src);
   ~Bureaucrat();
 
@@ -21,7 +21,7 @@ public:
   Bureaucrat &operator--();
 
   std::string const &getName(void) const;
-  size_t const &getGrade(void) const;
+  int const &getGrade(void) const;
 
   class GradeTooHighException : public std::exception
   {
